mcp_discord/requirements.txt
mcp>=1.0.0
discord.py>=2.3.0
python-dotenv>=1.0.0
